+++ src/main/java/com/github/hfantin/veiculos/infrastructure/web/dto/BrandResponse.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalDateTime;
 
 public class BrandResponse {
-    private Long id;
+    private Integer id;
     private String name;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
@@ -16,15 +16,15 @@ public class BrandResponse {
     // Constructors
     public BrandResponse() {}
 
-    public BrandResponse(Long id, String name, LocalDateTime createdAt) {
+    public BrandResponse(Integer id, String name, LocalDateTime createdAt) {
         this.id = id;
         this.name = name;
         this.createdAt = createdAt;
     }
 
     // Getters and Setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
